# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/mine/review/project/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "项目信息删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stru.IdParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SaveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mine/review/project/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "获取项目详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "name": "ids[]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ReviewProjectDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mine/review/project/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "获取项目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "modeCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "showStatus",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ReviewProjectListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mine/review/project/save": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "项目信息存储",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stru.ReviewProjectSaveParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SaveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mine/review/project/upStatus": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "项目状态更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stru.StatusParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SaveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ReviewProjectData": {
            "type": "object",
            "properties": {
                "createdDate": {
                    "type": "string"
                },
                "createdUser": {
                    "type": "string"
                },
                "deleted": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modeCode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedDate": {
                    "type": "string"
                },
                "updatedUser": {
                    "type": "string"
                }
            }
        },
        "api.ReviewProjectDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ReviewProjectData"
                    }
                }
            }
        },
        "api.ReviewProjectListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ReviewProjectData"
                    }
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.SaveResponse": {
            "type": "object",
            "properties": {
                "rowsAffected": {
                    "type": "integer"
                }
            }
        },
        "handler.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "stru.IdParam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "stru.ReviewProjectSaveData": {
            "type": "object",
            "required": [
                "modeCode",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modeCode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "stru.ReviewProjectSaveParam": {
            "type": "object",
            "required": [
                "action",
                "data"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/stru.ReviewProjectSaveData"
                }
            }
        },
        "stru.StatusParam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}